+++ packages/core/tests/CacheManager.test.ts
@@ -22,4 +22,33 @@ describe('CacheManager', () => {
     expect(manager.get('foo', { theme: 'dark' })).toBeNull();
     expect(manager.get('foo', { theme: 'dark', dir: 'rtl' })).toEqual({ value: 6 });
   });
+
+  it('clears all caches if no filter provided', () => {
+    manager.set('foo', { value: 1 }, { dir: 'ltr' });
+    manager.set('foo', { value: 2 }, { dir: 'ltr' });
+    manager.set('foo', { value: 3 }, { dir: 'ltr' });
+    manager.set('bar', { value: 1 }, { theme: 'light' });
+
+    // @ts-ignore Allow access
+    expect(manager.cache.size).toBe(2);
+
+    manager.clear();
+
+    // @ts-ignore Allow access
+    expect(manager.cache.size).toBe(0);
+  });
+
+  it('clears cache with the defined filter', () => {
+    manager.set('foo', { value: 1 }, { dir: 'ltr', theme: 'light' });
+    manager.set('foo', { value: 2 }, { dir: 'ltr', theme: 'dark' });
+    manager.set('foo', { value: 3 }, { dir: 'ltr', theme: 'light' });
+
+    manager.clear(unit => unit.theme === 'light');
+
+    // @ts-ignore Allow access
+    const cache = manager.cache.get('foo')!;
+
+    expect(cache).toHaveLength(1);
+    expect(cache[0]).toEqual({ dir: 'ltr', theme: 'dark', value: { value: 2 } });
+  });
 });